[FEAT] 모든 기능 구현

+++ src/main/java/lotto/view/DosInputView.java
@@ -27,7 +27,7 @@ public final class DosInputView implements InputView {
     }
 
     @Override
-    public int inputMoney() {
+    public long inputMoney() {
         return inputNumber(Text.INPUT_MONEY);
     }
