+++ src/test/java/lotto/view/FakeInputView.java
@@ -1,15 +1,15 @@
 package lotto.view;
 
 public class FakeInputView implements InputView {
-    private final int intMoney;
+    private final long intMoney;
     private final String prizeNumbers;
-    public FakeInputView(int intMoney, String prizeNumbers) {
+    public FakeInputView(long intMoney, String prizeNumbers) {
         this.intMoney = intMoney;
         this.prizeNumbers = prizeNumbers;
     }
 
     @Override
-    public int inputMoney() {
+    public long inputMoney() {
         return intMoney;
     }
